Item 1) on issue #2124
